Fix #40 (print fields ordered by tag number). (#67)
Also split out text_format_test's .proto file rather than sharing it
with canonical_test.

Also changed the "field name" stored in Data.ProtoLens.Message.FieldDescriptor
to be the name of that type, which is more consistent with how TextFormat
prints/parses such messages.  This change simplifies the new code.

Note: PR #48 previously referenced this bug, but it looks like that was a
typo since that fix was for string escaping, not field ordering.