Support OverloadedLabels. (#65)
* Support OverloadedLabels.

Adds a new package `lens-labels` (replacing `Data.ProtoLens.Field`)
which supports a newtype'd version of lenses, along with a typeclass
`HasLens` for using them with OverloadedLabels.

Note: In some cases it seems this causes problems with the lens-family
package and GHC's RankNTypes type inference, it seems because lens-family makes
setter functions like (.~) into higher-rank functions than are strictly
necessary.  Compare for example:
- http://hackage.haskell.org/package/lens-family-1.2.1/docs/Lens-Family2.html#v:.-126-
- http://hackage.haskell.org/package/lens-family-core-1.2.1/docs/Lens-Family.html#v:.-126-

Luckily, we can work around this by switching to a different package such as
`lens-family-core`, `lens` or `microlens`.