Add Data.ProtoLens.Any for storing arbitrary Messages. (#88)
Fixes #22.

Provides simple `pack` and `unpack` functions.  Since the `unpack` operation
is partial (and has multiple types of "failure"), I wasn't sure how to fit it
into a `Lens`; so I left that alone for the time being.

I turned this into a minor version bump.  It does add a new field to
`MessageDescriptor` which usually requires a major bump; but since that field
is internal to codegen, it shouldn't break any individual users.