Separate types into their own module, apart from field lenses. (#100)
Each proto file `foo/bar.proto` now generates two Haskell modules: `Foo.Bar` and `Foo.Bar'Fields`.  The former contains the proto types and `HasLens` instances; the latter contains the overloaded lens functions.  (In particular, the latter is not needed if you're using `OverloadedLabels`.)

This is a backwards-incompatible change due to the .cabal changes and module imports.  The actual code using the lenses should not need to change.