Consolidate proto-lens-descriptors into proto-lens. (#140)
There's no real reason for the two packages to be separate,
except that it increases the library size of `proto-lens`.  But since
`proto-lens-descriptors` is usually a transitive dependency anyway (through
`proto-lens-protoc`) I don't think it matters much.

Also remove an old bootstrapping-related flag that's not
used anymore.  (It should have been removed in #138.)

Helps address #109.

Ideally we'd merge `proto-lens-protoc` into the `proto-lens` package as well,
but I'm not sure yet how to do that without making bootstrapping more
complicated.