Fix the build on lts-12. (#213)
* Update changelogs for next 0.3.* release.

For:
 - proto-lens-protoc-0.3.1.1
 - proto-lens-protobuf-types-0.3.0.2

* Fix build of proto-lens-0.3.* on lts-12. (#211)

Also update `stack.yaml` to `lts-12.6` rather than a nightly.

The proto-lens-protoc-0.3.1.1 release had an upper bound
on `temporary` that was too restrictive.  Increase it,
and bump the version to `0.3.1.2`.

Also fix warnings and improve CI.  Turns out the Travis script wasn't robust enough and #195 let through
some warnings.  I fixed the script and confirmed it now catches such
issues.

Also bumped the tutorial and bootstrap scripts to the same LTS
as the main build.