Generated groups (#283)
* Make the decisions about map fields in collectDefinitions.

Previously Generate* had some of the responsibility, which
complicated the code (e.g., requiring passing `Env` to more
functions).

For efficiency, when looking for a map entry type we only look
in the immediate children (rather than transitively to all
nested types).  This required using a Tree as an intermediate
datastructure, but it didn't seem to complicate the code much.

* Support groups in the generated encodings.

I changed the `Definitions` module to keep track of the field number
of each group message type.  It meant a little more bookkeeping, but
simpified the `Generate.Encodings` module.  (Similar to #281, it
helps avoid the need to access an `Env` when generating the encodings.)