Remove the Encoding.Reflected API. (#293)
Also renames `unfinished{Parse,Build}Message` to `{parse,build}Message`.

Also gets rid of the remaining Reflected-related code. It also
changes WireValue from a GADT to a plain ADT.  The functionality
of the GADT isn't as useful for the generated encodings.

NOTE: this fixes #125 as a result of switching to the new implementation.  If
fields have the wrong wire type, we now store them in `unknownFields` rather
than failing the parse.  I updated a few unit tests that expected the old
behavior.

The only other unit test change that I made was to text_format_test, which
had a test that relied on the order in which fields are serialized to
the wire.  Previously, it was in order of tag type; now, it is (implicitly)
based on the order in the `.proto` file format.

Also updates the changelog and reruns bootstrapping.