Add `methodOptions` to `HasMethodImpl` to provide custom method options. (#331)
We provide the whole proto message, rather than pulling off individual fields. The set of fields may change for different builds of proto-lens (or based on extensions, though we can't see those directly due to #27).

Internally, we save the proto message in the source file as an encoded string, and decode it into a Haskell type lazily.